Made clear-cache button's label clickable.
