+++ src/components/Settings.tsx
@@ -5,7 +5,8 @@ import intl from 'react-intl-universal';
 import AltContainer from './AltContainer';
 import { useI18n } from '../context/I18nContext';
 import ClearConversationButton from './ClearConversationButton';
-import { useMasdifStatus } from '../context/MasdifClientContext';
+import { useMasdifClient, useMasdifStatus } from '../context/MasdifClientContext';
+import { useConversationContext } from '../context/ConversationContext';
 
 const InputGroup = styled.div`
     display: inline-block;
@@ -129,9 +130,41 @@ const getSettingsDescription = (setting: keyof SettingsValue) => {
 export type SettingsProps = {};
 
 export default function Settings(_: SettingsProps) {
+    const masdifClient = useMasdifClient();
     const masdifStatus = useMasdifStatus();
+    const [, convoDispatch] = useConversationContext();
     const [settings, setSettings] = useSettings();
     const [i18n, setI18n] = useI18n();
+
+    const clearConversation = async () => {
+        if (masdifClient && masdifStatus) {
+            // Clear messages and message feedback info.
+            convoDispatch({ type: 'CLEAR_CONVERSATION' });
+
+            // Current conversationId is overwritten.
+            const conversationId = await masdifClient.createConversation();
+            convoDispatch({ type: 'SET_CONVERSATION_ID', conversationId });
+
+            // New conversation is started with MOTD.
+            const info = await masdifClient.info(conversationId);
+            info.motd.reduce(
+                (p, text) =>
+                    p.then(
+                        () =>
+                            new Promise<void>(resolve => {
+                                convoDispatch({ type: 'DELAY_MOTD_RESPONSE' });
+                                window.setTimeout(() => {
+                                    convoDispatch({ type: 'ADD_RESPONSE', text });
+                                    resolve();
+                                }, 1000);
+                            }),
+                    ),
+                Promise.resolve(),
+            );
+            console.log('Conversation deleted.');
+        }
+    };
+
     return (
         <AltContainer>
             <SettingsContainer>
@@ -204,12 +237,12 @@ export default function Settings(_: SettingsProps) {
                     }
                 })}
                 <ButtonContainer
-                    title={intl.get('CHAT_SERVER_DOWN_TOOLTIP')}
+                    title={!masdifStatus ? intl.get('CHAT_SERVER_DOWN_TOOLTIP') : ''}
                     key={'clear-convo-group'}
                     $disabled={!masdifStatus}
                 >
-                    <ClearConversationButton key={'clear-convo-btn'} />
-                    <label key={'clear-convo-label'} htmlFor='clear-convo'>
+                    <ClearConversationButton key={'clear-convo-btn'} onClick={clearConversation} />
+                    <label key={'clear-convo-label'} htmlFor='clear-convo' onClick={clearConversation}>
                         {intl.get('SETTINGS_DESCRIPTION_CLEAR_CONVERSATION')}
                     </label>
                 </ButtonContainer>
